chore: move sys.version_info things to _compat
