+++ src/hist/plot.py
@@ -8,7 +8,7 @@ import numpy as np
 
 import hist
 
-from .typing import Literal
+from ._compat.typing import Literal
 
 try:
     import matplotlib.axes
